refactor(job_service): extracted html logic to web_service
